refactor(native-filters): refactor code of native filters (#12889)
* refactor(native-filters): refactor code of native filters

* refactor: update refactor dependencies

* refactor: update refactor dependencies

* lint: fix lint

* fix: merge with master

* chore: fix selector
