+++ superset-frontend/src/dashboard/actions/nativeFilters.ts
@@ -20,12 +20,12 @@
 import { ExtraFormData, makeApi } from '@superset-ui/core';
 import { Dispatch } from 'redux';
 import {
-  CurrentFilterState,
   Filter,
   FilterConfiguration,
-  SelectedValues,
 } from 'src/dashboard/components/nativeFilters/types';
 import { dashboardInfoChanged } from './dashboardInfo';
+import { CurrentFilterState } from '../reducers/types';
+import { SelectedValues } from '../components/nativeFilters/FilterConfigModal/types';
 
 export const SET_FILTER_CONFIG_BEGIN = 'SET_FILTER_CONFIG_BEGIN';
 export interface SetFilterConfigBegin {
